codegen/chart: fix parent key rendering for sidecars
This fixes a bug in how the sidecar values are rendered when
the sidecar name is not camel cased in model.Operator. It
camel cases the parent key for the sidecar similar to how
the sidecar name is rendered elsewhere.

For e.g., if the sidecar is named `foo-bar`, the rendering
in values.yaml looked as follows:
```
sidecar:
    foo-bar:
```
and the rendering in templates is as follows:
```
{{- $fooBar := $.Values.fooBarOperator.sidecars.fooBar }}
```

The actual container name cannot be camel cased in the Deployment
due to k8s requirement of a lowercase RFC 1123 name.